requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiofiles>=23.2.1
httpx>=0.27.0
python-dotenv>=1.0.1
rich>=13.7.0
dnspython>=2.6.0
pydantic>=2.0.0
pyarrow>=15.0.0
requests>=2.31.0
